pyproject.toml
[project]
name = "chm2pdf"
version = "2.0.0"
description = "Convert CHM (Compiled HTML Help) files to PDF with full content and style fidelity"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "weasyprint>=60.0",
    "pypdf>=3.0",
]

[project.optional-dependencies]
chm = ["pychm>=0.8"]

[project.scripts]
chm2pdf = "chm2pdf.cli:main"
chm2pdf-gui = "chm2pdf.gui:main"
